HW3. Servlet jsp+jstl

+++ firstapp_jsp/src/main/java/ru/geekbrains/servlet/Product.java
@@ -17,13 +17,13 @@ public class Product {
 
     private String description;
 
-    private BigDecimal price;
+    private Integer price;
 
     public Product(){
 
     }
 
-    public Product(Long id, String name, String description, BigDecimal price) {
+    public Product(Long id, String name, String description, Integer price) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -54,11 +54,11 @@ public class Product {
         this.description = description;
     }
 
-    public BigDecimal getPrice() {
+    public Integer getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(Integer price) {
         this.price = price;
     }
 }
